effective_preprocessing_pipelines/requirements.txt
certifi==2019.9.11
chardet==3.0.4
decorator==4.4.0
future==0.18.0
idna==2.8
imbalanced-learn==0.5.0
imblearn==0.0
joblib==0.14.0
liac-arff==2.4.0
networkx==2.3
numpy==1.17.2
openml==0.9.0
pandas==0.25.1
prettytable==0.7.2
psutil==5.6.3
python-dateutil==2.8.0
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
tqdm==4.36.1
urllib3==1.25.6
virtualenv==16.7.6
xmltodict==0.12.0
yellowbrick==1.0.1

